+++ Allocator.h
@@ -33,7 +33,7 @@ namespace hz
         StaticAllocation* allocate_static(Register = DC, bool = false);
 
         // `bool` parameter specifies whether the memory should be automatically freed at destruction
-        DynamicAllocation* allocate_dynamic(std::uint16_t = 1, bool = true);
+        DynamicAllocation* allocate_dynamic(std::uint32_t = 1, bool = true);
     };
 
     extern Allocator* _allocator;
