+++ AssemblerLinker.h
@@ -19,12 +19,12 @@ namespace hz
 		AssemblerParser* assembler_parser;
 
 	public:
-		std::uint16_t approximate_size()
+		std::uint32_t approximate_size()
 		{
 			// This is a very rough estimation of the number of bytes required by an inline
 			// assembly block that doesn't utilize .org directives to jump around. In that case,
 			// our linking doesn't really work. For now I am fine with this limitation.
-			return 3 * static_cast<std::uint16_t>(commands.size());
+			return 3 * static_cast<std::uint32_t>(commands.size());
 		}
 
 	public:
@@ -55,7 +55,7 @@ namespace hz
 	public:
 		virtual LinkerType ltype() const final override;
 		virtual bool optimize() final override;
-		virtual std::vector<InstructionCommand*> link(std::uint16_t) final override;
+		virtual std::vector<InstructionCommand*> link(std::uint32_t) final override;
 	};
 }
