+++ CommonToolchain.h
@@ -11,7 +11,7 @@
 namespace hz
 {
 	// shared between compiler and assembler
-	std::vector<InstructionCommand*> common_link(std::uint16_t = HALF_DWORD_MAX);
+	std::vector<InstructionCommand*> common_link(std::uint32_t = HALF_DWORD_MAX);
 	std::vector<std::uint8_t> common_emit(std::vector<InstructionCommand*>&&, const std::string& filepath);
 	void common_finalize(std::vector<std::uint8_t>&&, const std::string&);
 }
