+++ CompilerLinker.h
@@ -22,7 +22,7 @@ namespace hz
 	public:   
 		virtual LinkerType ltype() const final override;
 		virtual bool optimize() final override;
-		virtual std::vector<InstructionCommand*> link(std::uint16_t) final override;
+		virtual std::vector<InstructionCommand*> link(std::uint32_t) final override;
 	};
 }
