+++ InstructionCommand.cpp
@@ -17,7 +17,7 @@ namespace hz
 		marked_for_deletion = false;
 	}
 
-	InstructionCommand::InstructionCommand(Opcode opcode, Register dst, Register src, std::uint8_t imm, std::uint16_t mem, std::string branch_target)
+	InstructionCommand::InstructionCommand(Opcode opcode, Register dst, Register src, std::uint8_t imm, std::uint32_t mem, std::string branch_target)
 		: opcode{ opcode }, dst{ dst }, src{ src }, imm{ imm }, mem{ mem }, marked_for_deletion{ false }, branch_target{ std::move(branch_target) }
 	{
 		embedded_object_code = {};
