+++ Linker.h
@@ -28,7 +28,7 @@ namespace hz
 	public:
 		virtual LinkerType ltype() const = 0;
 		virtual bool optimize() = 0;
-		virtual std::vector<InstructionCommand*> link(std::uint16_t) = 0;
+		virtual std::vector<InstructionCommand*> link(std::uint32_t) = 0;
 	};
 
 	extern Linker* _linker;
